Create README.md
READ ME | This project is a market management system developed in the C programming language, using data structures such as Binary Search Trees (BST), Linked Lists, and QuickSort sorting. It was created as part of an academic project focused on file manipulation, efficient data organization, and simulation of a real-world sales environment.